+++ client/src/app/services/macros.service.ts
@@ -1,19 +1,25 @@
 import { Injectable } from '@angular/core';
 import { Select, Selector, Store } from '@ngxs/store';
-import { Observable } from 'rxjs';
+import { combineLatest, interval, Observable } from 'rxjs';
+import { switchMap } from 'rxjs/operators';
 
-import { IGame, IMacro, IMacroContainer } from '../../interfaces';
+import { ICharacter, IGame, IMacro, IMacroContainer, IPlayer } from '../../interfaces';
 import { GameState, MacrosState, SetActiveMacro, SetCurrentCommand, SettingsState } from '../../stores';
 import { GameService } from './game.service';
 
 import * as allMacros from '../../assets/content/_output/macros.json';
+import { OptionsService } from './options.service';
 @Injectable({
   providedIn: 'root'
 })
 export class MacrosService {
 
+  @Select(GameState.player) private player$: Observable<IPlayer>;
+  @Select(GameState.inGame) private inGame$: Observable<boolean>;
+  @Select(GameState.currentTarget) private currentTarget$: Observable<ICharacter>;
   @Select(SettingsState.activeWindow) private activeWindow$: Observable<string>;
   @Select(MacrosState.customMacros) private customMacros$: Observable<Record<string, IMacro>>;
+  @Select(MacrosService.currentPlayerActiveMacro) private activeMacro$: Observable<IMacro>;
 
   private macroMap: Record<string, IMacro> = {};
 
@@ -47,6 +53,7 @@ export class MacrosService {
 
   constructor(
     private store: Store,
+    private optionsService: OptionsService,
     private gameService: GameService
   ) {}
 
@@ -55,6 +62,7 @@ export class MacrosService {
     this.customMacros$.subscribe(c => this.parseMacroMap(c));
 
     this.watchForMacros();
+    this.autoAttackLoop();
   }
 
   public getMacroMatching(key: string): IMacro {
@@ -145,6 +153,17 @@ export class MacrosService {
     };
 
     document.addEventListener('keydown', macroListener);
+  }
+
+  private autoAttackLoop() {
+    interval(1000)
+      .pipe(switchMap(() => combineLatest([this.inGame$, this.player$, this.activeMacro$, this.currentTarget$])))
+      .subscribe(([inGame, player, macro, target]) => {
+        if (!inGame || !macro || !target || !this.optionsService.autoAttack || macro.ignoreAutoattackOption) return;
+        if (this.gameService.hostilityLevelFor(player, target) !== 'hostile') return;
+
+        this.gameService.sendCommandString(macro.macro, target.uuid);
+      });
   }
 
 }