+++ server/src/helpers/character/index.ts
@@ -2,6 +2,7 @@
 export * from './CalculatorHelper';
 export * from './CharacterHelper';
 export * from './CombatHelper';
+export * from './DailyHelper';
 export * from './DamageHelperMagic';
 export * from './DamageHelperOnesided';
 export * from './DamageHelperPhysical';