+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## [1.1.2](https://github.com/LandOfTheRair/lotr2/compare/v1.1.1...v1.1.2) (2020-11-25)
+
+
+### Bug Fixes
+
+* **ai:** monsters can attack other monsters again if they're always hostile ([3833b7a](https://github.com/LandOfTheRair/lotr2/commit/3833b7ae5c13da59cfe9bc6e56db0ea6df426e8b))
+* **ui:** command line will not show if it was hidden before when relaunching game ([7d40556](https://github.com/LandOfTheRair/lotr2/commit/7d405567d25717e8d748a917cb0c30959fa8a08c))
+* **ui:** italicize miss/block messages ([3fcf167](https://github.com/LandOfTheRair/lotr2/commit/3fcf1671d835f33650dc3e22df0dea7fa5df7504))
+* **ui:** should not auto-attack npcs you aren't hostile to ([8db9f8e](https://github.com/LandOfTheRair/lotr2/commit/8db9f8e8658ea185cb2ac820538f83cdf014c881))
+
+
+### Features
+
+* **npc:** daily items can now be bought and tracked properly ([f19c330](https://github.com/LandOfTheRair/lotr2/commit/f19c33010a12cd5d1cba28e9388b59e2d1bdcf3c))
+* **ui:** implement all options ([7dfab4a](https://github.com/LandOfTheRair/lotr2/commit/7dfab4aaa4f7fcac96af30cc0c823142ca755f0a)), closes [#42](https://github.com/LandOfTheRair/lotr2/issues/42)
+* **ui:** sound options are added ([126477b](https://github.com/LandOfTheRair/lotr2/commit/126477b41f26a16d91b886be476e4d117a9eb3d6)), closes [#42](https://github.com/LandOfTheRair/lotr2/issues/42)
+
+
+
 ## [1.1.1](https://github.com/LandOfTheRair/lotr2/compare/v1.1.0...v1.1.1) (2020-11-25)